Split some of the unit tests into separate test classes
