+++ tests/PhpSpreadsheetTests/Calculation/CalculationTest.php
@@ -58,73 +58,6 @@ class CalculationTest extends TestCase
         return require 'tests/data/CalculationBinaryComparisonOperation.php';
     }
 
-    /**
-     * @dataProvider providerGetFunctions
-     *
-     * @param string $category
-     * @param array|string $functionCall
-     * @param string $argumentCount
-     */
-    public function testGetFunctions($category, $functionCall, $argumentCount): void
-    {
-        self::assertIsCallable($functionCall);
-    }
-
-    public function providerGetFunctions(): array
-    {
-        return Calculation::getInstance()->getFunctions();
-    }
-
-    public function testIsImplemented(): void
-    {
-        $calculation = Calculation::getInstance();
-        self::assertFalse($calculation->isImplemented('non-existing-function'));
-        self::assertFalse($calculation->isImplemented('AREAS'));
-        self::assertTrue($calculation->isImplemented('coUNt'));
-        self::assertTrue($calculation->isImplemented('abs'));
-    }
-
-    /**
-     * @dataProvider providerCanLoadAllSupportedLocales
-     *
-     * @param string $locale
-     */
-    public function testCanLoadAllSupportedLocales($locale): void
-    {
-        $calculation = Calculation::getInstance();
-        self::assertTrue($calculation->setLocale($locale));
-    }
-
-    public function testInvalidLocaleReturnsFalse(): void
-    {
-        $calculation = Calculation::getInstance();
-        self::assertFalse($calculation->setLocale('xx'));
-    }
-
-    public function providerCanLoadAllSupportedLocales(): array
-    {
-        return [
-            ['bg'],
-            ['cs'],
-            ['da'],
-            ['de'],
-            ['en_us'],
-            ['es'],
-            ['fi'],
-            ['fr'],
-            ['hu'],
-            ['it'],
-            ['nl'],
-            ['nb'],
-            ['pl'],
-            ['pt'],
-            ['pt_br'],
-            ['ru'],
-            ['sv'],
-            ['tr'],
-        ];
-    }
-
     public function testDoesHandleXlfnFunctions(): void
     {
         $calculation = Calculation::getInstance();
@@ -425,19 +358,5 @@ class CalculationTest extends TestCase
     public function dataProviderBranchPruningFullExecution(): array
     {
         return require 'tests/data/Calculation/Calculation.php';
-    }
-
-    public function testUnknownFunction(): void
-    {
-        $workbook = new Spreadsheet();
-        $sheet = $workbook->getActiveSheet();
-        $sheet->setCellValue('A1', '=gzorg()');
-        $sheet->setCellValue('A2', '=mode.gzorg(1)');
-        $sheet->setCellValue('A3', '=gzorg(1,2)');
-        $sheet->setCellValue('A4', '=3+IF(gzorg(),1,2)');
-        self::assertEquals('#NAME?', $sheet->getCell('A1')->getCalculatedValue());
-        self::assertEquals('#NAME?', $sheet->getCell('A2')->getCalculatedValue());
-        self::assertEquals('#NAME?', $sheet->getCell('A3')->getCalculatedValue());
-        self::assertEquals('#NAME?', $sheet->getCell('A4')->getCalculatedValue());
     }
 }
